Add files via upload
Modified the MATLAB code to make use of MATLAB's built in function "angle2dcm()" which produces the rotation matrix based on yaw, pitch and roll and additionally allows you to specify the rotation sequence as 'XYZ', 'ZYX' etc. This is much more efficient than putting in our own rotation matrix function. Re-compiled the html folder with the latest changes.