Add files via upload
Changes made to the code to add prompts for user inputs. The ComputeDiodeResponse.m code now asks the user for initial roll, pitch and yaw conditions and which orientation parameter to step. Basic UX interface. The UserInterface code will allow in future for the ComputeDiodeResponse.m and other codes to have nice GUI interfaces for doing this.